move python files to resources module and misc fixes
