+++ python/pyspark/worker.py
@@ -25,7 +25,11 @@ import time
 # 'resource' is a Unix specific module.
 has_resource_module = True
 try:
-    import resource
+    from resource import RLIMIT_AS
+    from resource import RLIM_INFINITY
+    from resource import getrlimit
+    from resource import setrlimit
+    from resource import error
 except ImportError:
     has_resource_module = False
 import traceback
@@ -478,21 +482,21 @@ def main(infile, outfile):
         # set up memory limits
         memory_limit_mb = int(os.environ.get('PYSPARK_EXECUTOR_MEMORY_MB', "-1"))
         if memory_limit_mb > 0 and has_resource_module:
-            total_memory = resource.RLIMIT_AS
+            total_memory = RLIMIT_AS
             try:
-                (soft_limit, hard_limit) = resource.getrlimit(total_memory)
+                (soft_limit, hard_limit) = getrlimit(total_memory)
                 msg = "Current mem limits: {0} of max {1}\n".format(soft_limit, hard_limit)
                 print(msg, file=sys.stderr)
 
                 # convert to bytes
                 new_limit = memory_limit_mb * 1024 * 1024
 
-                if soft_limit == resource.RLIM_INFINITY or new_limit < soft_limit:
+                if soft_limit == RLIM_INFINITY or new_limit < soft_limit:
                     msg = "Setting mem limits to {0} of max {1}\n".format(new_limit, new_limit)
                     print(msg, file=sys.stderr)
-                    resource.setrlimit(total_memory, (new_limit, new_limit))
+                    setrlimit(total_memory, (new_limit, new_limit))
 
-            except (resource.error, OSError, ValueError) as e:
+            except (error, OSError, ValueError) as e:
                 # not all systems support resource limits, so warn instead of failing
                 print("WARN: Failed to set memory limit: {0}\n".format(e), file=sys.stderr)
